// src/core/enums/goal-intruments.enum.ts
export enum GOAL_INSTRUMENTS {
  ADCAP_AHORRO_PESOS = "AdCap Ahorro Pesos FCI - Clase A",
  ADCAP_PESOS_PLUS = "AdCap Pesos Plus FCI - Clase A",
  ADCAP_RENTA_FIJA = "AdCap Renta Fija Arg FCI - Clase A",
  ADCAP_RETORNO_TOTAL = "AdCap Retorno Total - Clase A",
  ADCAP_WISE = "AdCap Wise  - Clase A",
  ADCAP_BALANCEADO_II = "AdCap Balanceado II - Clase A",
  ADCAP_BALANCEADO = "AdCap Balanceado - Clase A", // CHEQUEAR Y AGREGAR A YML 
}